Updated to better match current usage
Hey theusaf! I shortened the incomplete message because of how it's currently being used. Removed "Please edit this when editing the page" because usually nobody changes it, the template is simply removed altogether after a few days without any edits. The new version should be clearer and easier to understand for newer users and it also works better with the refreshed {{incomplete}} and {{incomplete transcript}} templates, which have different text than before as they've been changed recently.